fix graph ranges
